move into new folder and update paths and amor notebook
