+++ src/ess/reflectometry/choppers/utils.py
